requirements.txt
asgiref==3.5.2
Django~=4.1.0
sqlparse==0.4.2
Pillow==9.2.0
social-auth-app-django==5.2.0
flower==1.2.0
celery==5.2.7
stripe==4.0.2
WeasyPrint==56.1
redis==4.3.4
django-import-export==3.2.0
whitenoise==6.4.0
psycopg2==2.9.3
uWSGI==2.0.21
daphne==3.0.2
